Change array to object
Change listbook arraylist instead Book object for store list of all book

+++ app/src/main/java/com/example/user/springbootwebserviceexample_java/Book.java
@@ -1,18 +1,18 @@
 package com.example.user.springbootwebserviceexample_java;
 
 public class Book {
-    long id;
+    int id;
     String name;
     String author;
 
-    public Book(long id, String name, String author) {
+    public Book(int id, String name, String author) {
         super();
         this.id = id;
         this.name = name;
         this.author = author;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
